+++ lib/services/services/database.dart
@@ -23,6 +23,24 @@ class DatabaseService {
     return response;
   }
 
+  // fetches data from table1 based on eq condition `columnValue`
+  // and inner joins on table2 (returns all rows) based on innerJoincolumn
+  static Future<PostgrestResponse> findApprovedEdits(String columnValue,
+      {String columnName = '${Constants.WORD_COLUMN}',
+      String innerJoincolumn = '${Constants.USER_EMAIL_COLUMN}',
+      String table1 = '${Constants.EDIT_HISTORY_TABLE}',
+      String table2 = '${Constants.USER_TABLE_NAME}',
+      bool sort = false}) async {
+    final response = await _supabase
+        .from(table1)
+        .select('*, $table2:$innerJoincolumn, $table2(*)')
+        .eq('$columnName', columnValue)
+        .eq('state', 'approved')
+        .order('${Constants.CREATED_AT_COLUMN}', ascending: sort)
+        .execute();
+    return response;
+  }
+
   static Future<PostgrestResponse> findRowByColumnValue(String columnValue,
       {String columnName = '${Constants.ID_COLUMN}',
       String tableName = '${Constants.VOCAB_TABLE_NAME}',