+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ### 0.6.0+23 July 21, 2023 (Unreleased)
 
 - Edit History is now transparent and public
+- Fix Responsiveness on desktop
+
 
 ### 0.5.9+22 July 21, 2023